opt4j problems
