--- MultiObj/src/NRPEvaluator.java
@@ -1,63 +1,0 @@
-
-import org.opt4j.core.problem.Evaluator;
-import org.opt4j.core.Objective.Sign;
-import org.opt4j.core.Objectives;
-
-/**
- * 
- * @author charl
- *
- *         Contains fitness functions for the two objectives:
- * 
- *         (1) Number of tests - min
- * 
- *         (2) Profit - max
- * 
- *         Both counts, not normalised to fall into range of 0 -1 Loaded into
- *         Objectives object with a label and indication to say if they should
- *         be minimised or maximised.
- * 
- */
-public class NRPEvaluator implements Evaluator<String> {
-
-	@Override
-	public Objectives evaluate(String phenotype) {
-		Objectives objectives = new Objectives();
-		objectives.add("TestSize", Sign.MIN, evaluateTestSizeFitness(phenotype));
-		objectives.add("Requirements Profit", Sign.MAX, evaluateFaultCovFitness(phenotype));
-		return objectives;
-	}
-
-	private int evaluateTestSizeFitness(String phenotype) {
-		int count = 0;
-		for (int i = 0; i < phenotype.length(); i++) {
-			if (phenotype.charAt(i) == '1')
-				count++;
-		}
-		return count;
-	}
-
-	public double evaluateFaultCovFitness(String phenotype) {
-		FM.readData(1000, 38);// Ugh! Do this once somewhere...
-		int faults = 38; // Again, should be injected
-		boolean[] faultsHit = new boolean[faults];
-		for (int i = 0; i < faults; i++)
-			faultsHit[i] = false;
-		for (int i = 0; i < phenotype.length(); i++) {
-			if (phenotype.charAt(i) == '1') {
-				// mark the faults found by the test
-				for (int j = 0; j < faults; j++) {
-					if (FM.faultMatrix[i][j] == 1)
-						faultsHit[j] = true;
-				}
-			}
-		}
-		// total collection of faults found
-		int count = 0;
-		for (int i = 0; i < faults; i++) {
-			if (faultsHit[i] == true)
-				count++;
-		}
-		return count;
-	}
-}
