--- MultiObj/src/NRPModule.java
@@ -1,8 +1,0 @@
-
-import org.opt4j.core.problem.ProblemModule;
-
-public class NRPModule extends ProblemModule {
-	protected void config() {
-		bindProblem(NRPCreator.class, NRPDecoder.class, NRPEvaluator.class);
-	}
-}
